refactor(#86): types 파일 분리
