+++ packages/ui/src/Button/ButtonText.tsx
@@ -1,12 +1,7 @@
-import type { ButtonSize } from '@/Button/Button';
-import { getTextStyles } from '@/Button/styles';
 import styled from '@emotion/styled';
 import type React from 'react';
-
-interface ButtonTextProps {
-  size: ButtonSize;
-  children: React.ReactNode;
-}
+import { getTextStyles } from './styles';
+import type { ButtonSize, ButtonTextProps } from './types';
 
 const ButtonText = ({ size, children }: ButtonTextProps) => {
   return <StyledText size={size}>{children}</StyledText>;
